vim-services/src/main/java/be/beanleaf/vim/utils/DateUtils.java
package be.beanleaf.vim.utils;

import java.time.LocalDateTime;
import java.time.LocalTime;
import java.time.ZoneId;
import java.util.Date;

public abstract class DateUtils {

  public static Date atEndOfMonth(Date date) {
    LocalDateTime localDateTime = dateToLocalDateTime(atEndOfDay(date));
    LocalDateTime endOfMonth = localDateTime
        .withDayOfMonth(localDateTime.getMonth().maxLength());
    return localDateTimeToDate(endOfMonth);
  }

  public static Date atStartOfDay(Date date) {
    LocalDateTime localDateTime = dateToLocalDateTime(date);
    LocalDateTime startOfDay = localDateTime.with(LocalTime.MIN);
    return localDateTimeToDate(startOfDay);
  }

  public static Date atEndOfDay(Date date) {
    LocalDateTime localDateTime = dateToLocalDateTime(date);
    LocalDateTime endOfDay = localDateTime.with(LocalTime.MAX);
    return localDateTimeToDate(endOfDay);
  }

  private static LocalDateTime dateToLocalDateTime(Date date) {
    return LocalDateTime.ofInstant(date.toInstant(), ZoneId.systemDefault());
  }

  public static Date localDateTimeToDate(LocalDateTime localDateTime) {
    return Date.from(localDateTime.atZone(ZoneId.systemDefault()).toInstant());
  }
}
